% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapfun.R
\name{bootstrapfun}
\alias{bootstrapfun}
\title{Function to perform traditional bootstrap for changepoint}
\usage{
bootstrapfun(X, n)
}
\arguments{
\item{X}{n x d matrix of pseudo-observations;}

\item{n}{length of the series.}
}
\value{
\item{cvm}{simulated value of the Cramer-von Mises statistic}

\item{ks}{simulated value of the Kolmogorov-Smirnov statistic}
}
\description{
This function select a bootstrap sample of indices and compute the Cramer-von Mises and Kolmogorov-Smirnov test statistics for changepoint
}
\keyword{internal}
