% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCopDStats.R
\name{cpCopDStats}
\alias{cpCopDStats}
\title{Function to compute changepoint statistics}
\usage{
cpCopDStats(U, n, d)
}
\arguments{
\item{U}{n x d matrix of pseudo-observations}

\item{n}{length of the series}

\item{d}{number of variables}
}
\value{
\item{statS}{Values of the Cramer-von Mises statistics}

\item{statT}{Values of the Kolmogorov-Smirnov statistics}
}
\description{
This function is used to compute the Cramer-von Mises and Kolmogorov-Smirnov test statistics for changepoint
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
