% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCopMultStats.R
\name{cpCopMultStats}
\alias{cpCopMultStats}
\title{Function to compute bootstrapped changepoint statistics}
\usage{
cpCopMultStats(MC, xi, s, n)
}
\arguments{
\item{MC}{matrix needed for multipliers}

\item{xi}{multipliers}

\item{s}{sequence of normalized values in (0,1)}

\item{n}{length of the series}
}
\value{
\item{statS}{Simulated values of the Cramer-von Mises statistics}

\item{statT}{Simulated values of the Kolmogorov-Smirnov statistics}
}
\description{
This function compute bootstrapped multipliers values for the Cramer-von Mises and Kolmogorov-Smirnov test statistics for changepoint
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
