% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCopMultStatsBKRSNonSeq.R
\name{cpCopMultStatsBKRSNonSeq}
\alias{cpCopMultStatsBKRSNonSeq}
\title{Function to compute the bootstrapped statistics for the BKRS process}
\usage{
cpCopMultStatsBKRSNonSeq(MC, MC1, grad, xi, s, n, d)
}
\arguments{
\item{MC}{matrix needed for multipliers}

\item{MC1}{matrices needed for multipliers}

\item{grad}{gradient of the copula}

\item{xi}{multipliers}

\item{s}{sequence of normalized values in (0,1)}

\item{n}{length of the series}

\item{d}{number of variables}
}
\value{
\item{statS}{Values of the Cramer-von Mises statistics}

\item{statT}{Values of the Kolmogorov-Smirnov statistics}
}
\description{
This function is used to compute bootstrapped statistics from the BKRS process in the non-sequential case
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
