% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCopMultStatsBKRSSeq.R
\name{cpCopMultStatsBKRSSeq}
\alias{cpCopMultStatsBKRSSeq}
\title{Function to compute the statistics for the BKRS process}
\usage{
cpCopMultStatsBKRSSeq(U, grad, xi, n, d)
}
\arguments{
\item{U}{matrix needed for multipliers}

\item{grad}{gradient of the copula}

\item{xi}{multipliers}

\item{n}{length of the series}

\item{d}{number of variables}
}
\value{
\item{statS}{Values of the Cramer-von Mises statistics}

\item{statT}{Values of the Kolmogorov-Smirnov statistics}
}
\description{
This function is used to compute bootstrapped statistics from the BKRS process in the sequential case
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
