% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCopStats.R
\name{cpCopStats}
\alias{cpCopStats}
\title{Function to compute the statistics for the traditional empirical process}
\usage{
cpCopStats(x)
}
\arguments{
\item{x}{n x d matrix of pseudo-observations}
}
\value{
\item{M}{Matrix for the cdf}

\item{statS}{Values of the Cramer-von Mises statistics}

\item{statT}{Values of the Kolmogorov-Smirnov statistics}
}
\description{
This function is used to compute the empirical process used in the changepoint tests
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
