% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpCopStatsBKRS.R
\name{cpCopStatsBKRS}
\alias{cpCopStatsBKRS}
\title{Function to compute the statistics for the BKRS process}
\usage{
cpCopStatsBKRS(x, n, d)
}
\arguments{
\item{x}{n x d matrix of pseudo-observations}

\item{n}{length of the series}

\item{d}{number of variables}
}
\value{
\item{MC}{Matrix needed for multipliers}

\item{MC1}{Matrices needed for multipliers}

\item{grad1}{Esatimated gradient of the copula}
}
\description{
This function is used to compute the empirical process used in the changepoint tests
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
