% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp.cdf.R
\name{emp.cdf}
\alias{emp.cdf}
\title{Function to compute the empirical cdf at given points}
\usage{
emp.cdf(x, u)
}
\arguments{
\item{x}{n x d matrix of pseudo-observations}

\item{u}{evaluation points of the cdf}
}
\value{
\item{M}{Values of the empirical process}

\item{cumsum}{Associated cumulativ sums}
}
\description{
This function is used to compute the empirical process used in the changepoint tests
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
