% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplierDNonSeqfun.R
\name{multiplierDNonSeqfun}
\alias{multiplierDNonSeqfun}
\title{Function to perform multiplier bootstrap for changepoint}
\usage{
multiplierDNonSeqfun(MC, MC1, grad, s, n, d)
}
\arguments{
\item{MC}{matrix needed for multipliers}

\item{MC1}{matrices needed for multipliers}

\item{grad}{gradient of the copula}

\item{s}{sequence of normalized values in (0,1)}

\item{n}{length of the series}

\item{d}{number of variables}
}
\value{
\item{cvm}{simulated value of the Cramer-von Mises statistic}

\item{ks}{simulated value of the Kolmogorov-Smirnov statistic}
}
\description{
This function simulates a random sample of Gaussian multipliers  null hypothesis of a Gaussian HMM
and compute the Cramer-von Mises and Kolmogorov-Smirnov test statistics.
}
\references{
Nasri, B. R. Remillard, B., & Bahraoui, T. (2022).
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
