% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplierDSeqfun.R
\name{multiplierDSeqfun}
\alias{multiplierDSeqfun}
\title{Function to perform multiplier bootstrap for changepoint}
\usage{
multiplierDSeqfun(U, grad, n, d)
}
\arguments{
\item{U}{matrix needed for multipliers}

\item{grad}{gradient of the copula}

\item{n}{length of the series}

\item{d}{number of variables.}
}
\value{
\item{cvm}{simulated value of the Cramer-von Mises statistic}

\item{ks}{simulated value of the Kolmogorov-Smirnov statistic}
}
\description{
This function simulates a random sample of Gaussian multipliers  null hypothesis of a Gaussian HMM
and compute the Cramer-von Mises and Kolmogorov-Smirnov test statistics.
}
\references{
Nasri, B. R. Remillard, B., & Bahraoui, T. (2022).
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
