% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplierfun.R
\name{multiplierfun}
\alias{multiplierfun}
\title{Function to perform multiplier bootstrap for changepoint}
\usage{
multiplierfun(MC, s, n)
}
\arguments{
\item{MC}{n x n matrix = MM  - C, with MM[i,j] = 1(Xi <= Xj) and C=mean(M[,j]);}

\item{n}{length of the series.}
}
\value{
\item{cvm}{simulated value of the Cramer-von Mises statistic}

\item{ks}{simulated value of the Kolmogorov-Smirnov statistic}
}
\description{
This function simulates a random sample of Gaussian multipliers  null hypothesis of a Gaussian HMM
and compute the Cramer-von Mises and Kolmogorov-Smirnov test statistics.
}
\references{
Chapter 8 of B. Remillard (2013). Statistical Methods for Financial Engineering,
Chapman and Hall/CRC Financial Mathematics Series, Taylor & Francis.
}
\author{
Bouchra R Nasri  and Bruno N Remillard, August 6, 2020
}
\keyword{internal}
