\name{pseudos}
\docType{data}
\alias{pseudos}
\title{Pseudo-observations}
\description{Pseudo-observations used in Nasri, Remillard, Bahraoui (2021). The values represent conditional cdfs of Gaussian HMM models applied to log-returns of Nasdaq and Dow Jones Industrial indexes from 2007 and 2008. If the models are correct, the pseudo-observations should be almost iid with uniform distribution.}
\format{Pseudo-observations from Gaussian HMM models with 3 regimes for log-returns of the to Nasdaq index and Dow Jones Industrial indexes from 2007 and 2008.
\itemize{
  \item 1st column: pseudo-observations of a Gaussian HMM model with 3 regimes applied to the Nasdaq log-returns.
  \item 2nd column: pseudo-observations of a Gaussian HMM model with 3 regimes applied to the Dow Jones Industrial log-returns. }
 }
 \usage{data(pseudos)}
 \keyword{datasets}

